+++ factcheck/internal/repo/message.go
@@ -6,7 +6,7 @@ import (
 	"github.com/jackc/pgx/v5/pgtype"
 
 	"github.com/kaogeek/line-fact-check/factcheck"
-	"github.com/kaogeek/line-fact-check/factcheck/postgres/sqlcgen"
+	"github.com/kaogeek/line-fact-check/factcheck/models/postgres"
 )
 
 // RepositoryMessage defines the interface for message data operations
